refactor: 타입스크립트 적용 및 ts, tsx 파일로 수정
