+++ src/App.tsx
@@ -1,11 +1,11 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
+import Header from "./components/Layout/Header";
 import HomePage from "./pages/HomePage/HomePage";
 import LoginPage from "./pages/LoginPage/LoginPage";
 import MarketPage from "./pages/MarketPage/MarketPage";
+import ItemPage from "./pages/ItemPage/ItemPage";
 import AddItemPage from "./pages/AddItemPage/AddItemPage";
 import CommunityFeedPage from "./pages/CommunityFeedPage/CommunityFeedPage";
-import Header from "./components/Layout/Header";
-import ItemPage from "./pages/ItemPage/ItemPage";
 
 function App() {
   return (
