+++ src/main.tsx
@@ -1,9 +1,9 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import App from './App.jsx';
+import App from './App.js';
 import { ThemeProvider } from 'styled-components';
-import theme from './styles/theme.js';
-import GlobalStyle from './styles/GlobalStyle.js';
+import theme from './styles/theme.ts';
+import GlobalStyle from './styles/GlobalStyle.ts';
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
