+++ src/pages/ItemPage/ItemPage.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import styled from 'styled-components';
 import { Container, LineDivider, StyledLink } from '../../styles/CommonStyles';
